archive working perfectely

+++ lib/services/firestore_db.dart
@@ -1,6 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:googlekeep/login_info.dart';
 import 'package:googlekeep/model/MyNoteModel.dart';
 import 'db.dart';
 
@@ -9,22 +10,26 @@ class FireDB {
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
   createNewNoteFirestore(Note note ) async {
-    final User? current_user = _auth.currentUser;
-    await FirebaseFirestore.instance
-        .collection("notes")
-        .doc(current_user!.email)
-        .collection("usernotes")
-        .doc(note.uniqueID)
-        .set({
-      "Title": note.title,
-      "content": note.content,
-      "uniqueID" : note.uniqueID,
-      "date": note.createdTime,
-      // "Title": "New test title",
-      //  "content": "test notes",
-      //  "date": DateTime.now(),
-    }).then((_) {
-      print("DATA ADDED SUCCESSFULLY");
+    LocalDataSaver.getSyncSet().then((isSyncOn) async{
+      if(isSyncOn.toString() == "true"){
+        final User? current_user = _auth.currentUser;
+        await FirebaseFirestore.instance
+            .collection("notes")
+            .doc(current_user!.email)
+            .collection("usernotes")
+            .doc(note.uniqueID)
+            .set({
+          "Title": note.title,
+          "content": note.content,
+          "uniqueID" : note.uniqueID,
+          "date": note.createdTime,
+          // "Title": "New test title",
+          //  "content": "test notes",
+          //  "date": DateTime.now(),
+        }).then((_) {
+          print("DATA ADDED SUCCESSFULLY");
+        });
+      }
     });
   }
 
@@ -52,28 +57,37 @@ class FireDB {
   }
 
   updateNoteFirestore(Note note) async {
-    final User? current_user = _auth.currentUser;
-    await FirebaseFirestore.instance
-        .collection("notes")
-        .doc(current_user!.email)
-        .collection("usernotes")
-        .doc(note.uniqueID.toString())
-        .update({"Title": note.title.toString(), "content": note.content}).then(
-            (_) {
-      print("DATA ADDED SUCCESFULLY");
+    LocalDataSaver.getSyncSet().then((isSyncOn) async{
+      if(isSyncOn.toString() == "true"){
+        final User? current_user = _auth.currentUser;
+        await FirebaseFirestore.instance
+            .collection("notes")
+            .doc(current_user!.email)
+            .collection("usernotes")
+            .doc(note.uniqueID.toString())
+            .update({"Title": note.title.toString(), "content": note.content}).then(
+                (_) {
+              print("DATA ADDED SUCCESFULLY");
+            });
+      }
     });
   }
 
   deleteNoteFirestore(Note note) async {
-    final User? current_user = _auth.currentUser;
-    await FirebaseFirestore.instance
-        .collection("notes")
-        .doc(current_user!.email.toString())
-        .collection("usernotes")
-        .doc(note.uniqueID.toString())
-        .delete()
-        .then((_) {
-      print("DATA DELETED SUCCESS FULLY");
+    LocalDataSaver.getSyncSet().then((isSyncOn) async{
+      if(isSyncOn.toString() == "true"){
+        final User? current_user = _auth.currentUser;
+        await FirebaseFirestore.instance
+            .collection("notes")
+            .doc(current_user!.email.toString())
+            .collection("usernotes")
+            .doc(note.uniqueID.toString())
+            .delete()
+            .then((_) {
+          print("DATA DELETED SUCCESS FULLY");
+        });
+      }
     });
+
   }
 }
